% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpResponse.R
\name{httpResponse.contentType}
\alias{httpResponse.contentType}
\title{Create a Content Type String}
\usage{
httpResponse.contentType(type = "text")
}
\arguments{
\item{type}{file type or standard extension}
}
\value{
A character string with the appropriate MIME type.
}
\description{
The \code{type} parameter is typically the file extension.
}
\examples{
httpResponse.contentType('text')
httpResponse.contentType('json')
httpResponse.contentType('png')
}
