% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpResponse.R
\name{httpResponse.header}
\alias{httpResponse.header}
\title{Create a HTTP Response Header}
\usage{
httpResponse.header(type = "text")
}
\arguments{
\item{type}{file type used to generate the Content-Type string}
}
\value{
A text string containing a valid HTTP Response header.
}
\description{
This function will generate a header string containing only a minimal
list of possible response header elements including:
\itemize{
\item{\code{Content-Type}}
}
}
\examples{
httpResponse.header('text')
httpResponse.header('json')
httpResponse.header('png')
}
