% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpResponse.R
\name{mimeType}
\alias{mimeType}
\title{Create a MIME Type String}
\usage{
mimeType(type = "text")
}
\arguments{
\item{type}{file type or standard extension}
}
\value{
A character string with the appropriate MIME type.
}
\description{
The \code{type} parameter is typically the file extension.
}
\examples{
mimeType('text')
mimeType('json')
mimeType('png')
}
\references{
\href{https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types/Complete_list_of_MIME_types}{MDN MIME types for the web}

\href{http://www.iana.org/assignments/media-types/media-types.xhtml}{IANA all media types}

\href{https://en.wikipedia.org/wiki/Media_type}{Wikipedia media types}
}
